% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{append_region}
\alias{append_region}
\alias{append_region,EpivizEnvironment-method}
\title{Generic method to add navigation regions}
\usage{
append_region(x, ...)

\S4method{append_region}{EpivizEnvironment}(x, chr, start, end, return_parent = FALSE)
}
\arguments{
\item{x}{an object of type \code{\link{EpivizEnvironment}}}

\item{...}{other parameters}

\item{chr}{chromosome id. ex. "chr11"}

\item{start}{genomic region start}

\item{end}{genomic region end}

\item{return_parent}{To return the parent or the new navigation element. Defaults to FALSE}
}
\description{
Generic method to add navigation regions
}
