% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{append_region}
\alias{append_region}
\title{Generic method to  add navigation regions}
\usage{
append_region(x, ...)
}
\arguments{
\item{x}{an object of type \code{\link{EpivizEnvironment}}}

\item{...}{other parameters - genomic locations -  chr, start and end.}
}
\description{
Generic method to  add navigation regions
}
