% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_data.R
\docType{data}
\name{bcode_eset}
\alias{bcode_eset}
\title{Example expression data from the Gene Expression barcode project}
\format{
An \code{Biobase::ExpressionSet} object
}
\usage{
data(bcode_eset)
}
\description{
See vignette \code{data_preprocessing} for code to build this object.
}
