% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{BlocksTrack}
\alias{BlocksTrack}
\alias{BlocksTrack,EpivizEnvironment-method}
\title{Method to add Blocks Track}
\usage{
BlocksTrack(x, y, ...)

\S4method{BlocksTrack}{EpivizEnvironment}(x, y, ...)
}
\arguments{
\item{x}{an object of type \code{\link{EpivizEnvironment}} or \code{\link{EpivizNavigation}}}

\item{y}{a genomic data object}

\item{...}{other parameters for the plot method}
}
\description{
Method to add Blocks Track
}
