% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_data.R
\docType{data}
\name{cgi_gr}
\alias{cgi_gr}
\title{Locations of CpG Islands}
\format{
A \code{GenomicRanges::GRanges} object with locations of CpG Islands in hg19
}
\usage{
data(cgi_gr)
}
\description{
Locations of CpG Islands
}
