% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chart_type_to_tag_name}
\alias{chart_type_to_tag_name}
\title{Get epiviz chart component from data object}
\usage{
chart_type_to_tag_name(ms_obj, chart)
}
\arguments{
\item{ms_obj}{data infer chart type from data object}

\item{chart}{explicitly define chart type}
}
\value{
epiviz chart component tag name
}
\description{
Get epiviz chart component from data object
}
