% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.constructURL}
\alias{.constructURL}
\title{Construct URL for Websocket connection between R and UI}
\usage{
.constructURL(host = "localhost", port = 7123L, path = "")
}
\arguments{
\item{host}{host}

\item{port}{port}

\item{path}{path}
}
\value{
url
}
\description{
Construct URL for Websocket connection between R and UI
}
