% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-inits.R
\name{.initialize_chart}
\alias{.initialize_chart}
\title{Initialize Epiviz Chart based on chart type}
\usage{
.initialize_chart(chart_type, ...)
}
\arguments{
\item{chart_type}{Chart type.}

\item{...}{Arguments for \code{\link{EpivizChart}} objects.}
}
\description{
Initialize Epiviz Chart based on chart type
}
