% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.register_all_the_epiviz_things}
\alias{.register_all_the_epiviz_things}
\title{(taken from epivizr) register epiviz actions}
\usage{
.register_all_the_epiviz_things(srv, app)
}
\arguments{
\item{srv}{epivizrServer object}

\item{app}{EpivizApp object}
}
\description{
(taken from epivizr) register epiviz actions
}
