% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-inits.R
\name{epiviz}
\alias{epiviz}
\title{Initialize an \code{\link[epivizrChart]{EpivizApp}} app.}
\usage{
epiviz(...)
}
\arguments{
\item{...}{Additional params to pass to \code{\link[epivizrServerChart]{EpivizServer}}}
}
\value{
An object of class \code{\link[epivizrChart]{EpivizApp}}
}
\description{
Initialize an \code{\link[epivizrChart]{EpivizApp}} app.
}
\examples{

}
