% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizApp-class.R
\docType{class}
\name{EpivizApp-class}
\alias{EpivizApp-class}
\alias{EpivizApp}
\title{Data container for an Epiviz App.}
\description{
Data container for an Epiviz App.
}
\section{Fields}{

\describe{
\item{\code{data_mgr}}{}
}}


