% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizCharts-classes.R
\docType{class}
\name{EpivizBlocksTrack-class}
\alias{EpivizBlocksTrack-class}
\alias{EpivizBlocksTrack}
\title{Data container for an Epiviz Blocks Track.}
\description{
Data container for an Epiviz Blocks Track.
}

