% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizChart-class.R
\docType{class}
\name{EpivizChart-class}
\alias{EpivizChart-class}
\alias{EpivizChart}
\title{Data container for an epiviz chart.}
\description{
Data container for an epiviz chart.
}
\section{Fields}{

\describe{
\item{\code{mgr}}{An object of class \code{\link[epivizrChart]{EpivizChartDataMgr}} used to serve data to epiviz chart.}

\item{\code{obj}}{An object of class \code{\link[epivizrData]{EpivizData}}.}
}}


