% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epivizChart.R
\name{epivizChart}
\alias{epivizChart}
\title{Create EpivizChart object.}
\usage{
epivizChart(data_object, datasource_name,
  datasource_origin_name = deparse(substitute(data_object)),
  epiviz_env = NULL, chart_type = NULL, chr = NULL, start = NULL,
  end = NULL, settings = NULL, colors = NULL, ...)
}
\arguments{
\item{data_object}{(EpivzData)}

\item{datasource_name}{(character)}

\item{datasource_origin_name}{(character)}

\item{chart_type}{(character) BlocksTrack, HeatmapPlot, LinePlot,LineTrack, ScatterPlot, StackedLinePlot, StackedLineTrack}

\item{chr}{(character)}

\item{start}{(numeric)}

\item{end}{(numeric)}

\item{settings}{(list) list of settings}

\item{colors}{(list) list of colors}

\item{...}{(type, columns, chr, start, end)}

\item{data_mgr}{(EpivizChartDataMgr)}
}
\value{
An object of class \code{\link[epivizrChart]{EpivizChart}}
}
\description{
Create EpivizChart object.
}
\examples{
# see package vignette for example usage

}
