% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizDataSource-class.R
\docType{class}
\name{EpivizDataSource-class}
\alias{EpivizDataSource-class}
\alias{EpivizDataSource}
\title{Data container for an Epiviz Data Source component.}
\description{
Data container for an Epiviz Data Source component.
}
\section{Fields}{

\describe{
\item{\code{provider_type}}{(character)}

\item{\code{provider_id}}{(character)}

\item{\code{provider_url}}{(character)}
}}

\section{Methods}{

\describe{
\item{\code{get_attributes()}}{Get attributes for rendering web component}

\item{\code{get_component_type()}}{Get component type for prefix of random id generator}

\item{\code{get_name()}}{Get name of Epiviz Web Component}

\item{\code{get_provider_id()}}{Get provider id}

\item{\code{get_provider_type()}}{Get provider type}

\item{\code{get_provider_url()}}{Get provider url}

\item{\code{render_component(shiny = FALSE)}}{Render to html}

\item{\code{set_provder_url(url)}}{Set provider url}

\item{\code{set_provider_id(id)}}{Set provider id}

\item{\code{set_provider_type(type)}}{Set provider type}
}}

