% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-inits.R
\name{epivizDS}
\alias{epivizDS}
\title{Initialize an \code{\link[epivizrChart]{EpivizDataSource}} object.}
\usage{
epivizDS(type, p_id, url, ...)
}
\arguments{
\item{type}{TODO}

\item{p_id}{TODO}

\item{url}{TODO}

\item{...}{Additional params to pass to \code{\link[epivizrChart]{EpivizDataSource}}}
}
\value{
An object of class \code{\link[epivizrChart]{EpivizDataSource}}
}
\description{
Initialize an \code{\link[epivizrChart]{EpivizDataSource}} object.
}
\examples{

}
