% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizEnvironment-class.R
\docType{class}
\name{EpivizEnvironment-class}
\alias{EpivizEnvironment-class}
\alias{EpivizEnvironment}
\title{Epiviz Environment Class}
\description{
Epiviz Environment Class
}
\section{Fields}{

\describe{
\item{\code{chr}}{(character) chromosome to to display in environment plot.}

\item{\code{start}}{(integer) start location to display in environment plot.}

\item{\code{end}}{(integer) end location to to display in environment plot.}
}}

\section{Methods}{

\describe{
\item{\code{add(polymer_obj)}}{Add chart or navigation to environment}
}}

