% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizCharts-classes.R
\docType{class}
\name{EpivizGenesTrack-class}
\alias{EpivizGenesTrack-class}
\alias{EpivizGenesTrack}
\title{Data container for an Epiviz Genes Track.}
\description{
Data container for an Epiviz Genes Track.
}

