% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizCharts-classes.R
\docType{class}
\name{EpivizHeatmapPlot-class}
\alias{EpivizHeatmapPlot-class}
\alias{EpivizHeatmapPlot}
\title{Data container for an Epiviz Heatmap Plot.}
\description{
Data container for an Epiviz Heatmap Plot.
}
\section{Methods}{

\describe{
\item{\code{get_chart_type()}}{Get chart type}

\item{\code{get_default_settings()}}{Get default settings}

\item{\code{get_name()}}{Get name}
}}

