% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizCharts-classes.R
\docType{class}
\name{EpivizLineTrack-class}
\alias{EpivizLineTrack-class}
\alias{EpivizLineTrack}
\title{Data container for an Epiviz Line Track.}
\description{
Data container for an Epiviz Line Track.
}

