% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epivizNavigation.R
\name{epivizNavigation}
\alias{epivizNavigation}
\title{Create an \code{\link[epivizrChart]{EpivizNavigation}} object.}
\usage{
epivizNavigation(chr = "chr11", start = 99800000, end = 103383180)
}
\arguments{
\item{chr}{(character) chromosome to browse to on app startup.}

\item{start}{(integer) start location to browse to on app startup.}

\item{end}{(integer) end location to browse to on app startup.}

\item{data_mgr}{(EpivizChartDataMgr)}
}
\value{
An object of class \code{\link[epivizrChart]{EpivizNavigation}}
}
\description{
Create an \code{\link[epivizrChart]{EpivizNavigation}} object.
}
\examples{
# see package vignette for example usage
epiviz <- epivizEnvironment(chr="chr11", start=99800000, end=103383180)

}
