% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epivizPlot.R
\name{epivizPlot}
\alias{epivizPlot}
\title{Create EpivizChart object.}
\usage{
epivizPlot(data_object, datasource_name,
  datasource_origin_name = deparse(substitute(data_object)),
  data_mgr = EpivizChartDataMgr$new(), chart_type = NULL, settings = NULL,
  colors = NULL, ...)
}
\arguments{
\item{data_object}{(EpivzData)}

\item{datasource_name}{(character)}

\item{datasource_origin_name}{(character)}

\item{data_mgr}{(EpivizChartDataMgr)}

\item{chart_type}{(character) BlocksTrack, HeatmapPlot, LinePlot,LineTrack, ScatterPlot, StackedLinePlot, StackedLineTrack}

\item{settings}{(list) list of settings}

\item{colors}{(list) list of colors}

\item{...}{(type, columns, chr, start, end)}
}
\value{
An object of class \code{\link[epivizrChart]{EpivizChart}}
}
\description{
Create EpivizChart object.
}
\examples{
# see package vignette for example usage

}
