% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizPolyDataMgr-class.R
\docType{class}
\name{EpivizPolyDataMgr-class}
\alias{EpivizPolyDataMgr-class}
\alias{EpivizPolyDataMgr}
\title{Class providing data manager for epiviz charts.}
\description{
Class providing data manager for epiviz charts.
}
\section{Fields}{

\describe{
\item{\code{.ms_list}}{(environment)}

\item{\code{.ms_idCounter}}{(integer)}
}}

\section{Methods}{

\describe{
\item{\code{add_measurements(obj, datasource_name = NULL,
  datasource_origin_name = deparse(substitute(obj)), ...)}}{register measurements in data manager}
}}

