% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizCharts-classes.R
\docType{class}
\name{EpivizScatterPlot-class}
\alias{EpivizScatterPlot-class}
\alias{EpivizScatterPlot}
\title{Data container for an Epiviz Scatter Plot.}
\description{
Data container for an Epiviz Scatter Plot.
}

