% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizCharts-classes.R
\docType{class}
\name{EpivizStackedLinePlot-class}
\alias{EpivizStackedLinePlot-class}
\alias{EpivizStackedLinePlot}
\title{Data container for an Epiviz Stacked Line Plot.}
\description{
Data container for an Epiviz Stacked Line Plot.
}

