% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizCharts-classes.R
\docType{class}
\name{EpivizTranscriptTrack-class}
\alias{EpivizTranscriptTrack-class}
\alias{EpivizTranscriptTrack}
\title{Data container for an Epiviz Transcript Track.}
\description{
Data container for an Epiviz Transcript Track.
}
\section{Methods}{

\describe{
\item{\code{get_component_type()}}{Get component type for prefix of random id generator}

\item{\code{get_default_colors()}}{Get default colors}

\item{\code{get_default_settings()}}{Get default settings}

\item{\code{get_name()}}{Get name of Epiviz Web Component}
}}

