% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizViewComponent-class.R
\docType{class}
\name{EpivizViewComponent-class}
\alias{EpivizViewComponent-class}
\alias{EpivizViewComponent}
\title{Data container for an Epiviz web component.}
\description{
Data container for an Epiviz web component.
}
\section{Fields}{

\describe{
\item{\code{chr}}{(CharacterOrNULL) Chromosome location.}

\item{\code{start}}{(NumericOrNULL) Start location.}

\item{\code{end}}{(NumericOrNULL) End location.}

\item{\code{measurements}}{(ListOrNULL) list of measurements of class \code{\link{EpivizMeasurement}}.}
}}

\section{Methods}{

\describe{
\item{\code{get_attributes()}}{Get attributes for rendering web component}

\item{\code{get_chr()}}{Get chromosome}

\item{\code{get_end()}}{Get end}

\item{\code{get_measurements()}}{Get measurements}

\item{\code{get_start()}}{Get start}

\item{\code{set_chr(chr)}}{Set the chromosome}

\item{\code{set_end(end)}}{Set end}

\item{\code{set_measurements(ms)}}{Set measurements}

\item{\code{set_start(start)}}{Set start}
}}

