% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizWebComponent-class.R
\docType{class}
\name{EpivizWebComponent-class}
\alias{EpivizWebComponent-class}
\alias{EpivizWebComponent}
\title{Data container for an Epiviz Web component.}
\description{
Data container for an Epiviz Web component.
}
\section{Fields}{

\describe{
\item{\code{data_mgr}}{\code{\link[epivizrChart]{EpivizChartDataMgr}}}

\item{\code{name}}{(character) Epiviz chart type (tag name).}

\item{\code{class}}{(CharacterOrNULL) Epiviz chart's class attribute.}

\item{\code{id}}{(character) Epiviz chart's id attribute.}
}}

\section{Methods}{

\describe{
\item{\code{get_attributes()}}{Get attributes for rendering web component}

\item{\code{get_class()}}{Get class}

\item{\code{get_data_mgr()}}{Get data manager}

\item{\code{get_id()}}{Get id}

\item{\code{get_name()}}{Get name of Epiviz Web Component}

\item{\code{set_class(class)}}{Set chart class}

\item{\code{set_id(id)}}{Set chart id}

\item{\code{set_name(name)}}{Set name}
}}

