% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_registered_data_types}
\alias{get_registered_data_types}
\title{Construct URL for Websocket connection between R and UI}
\usage{
get_registered_data_types()
}
\value{
url
}
\description{
Construct URL for Websocket connection between R and UI
}
