% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{GuideTrack}
\alias{GuideTrack}
\alias{GuideTrack,EpivizEnvironment-method}
\title{Method to add Guide Track}
\usage{
GuideTrack(x, y, ...)

\S4method{GuideTrack}{EpivizEnvironment}(x, y, ...)
}
\arguments{
\item{x}{an object of type \code{\link{EpivizEnvironment}} or \code{\link{EpivizNavigation}}}

\item{y}{a genomic data object}

\item{...}{other parameters for the plot method}
}
\description{
Method to add Guide Track
}
