% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{HeatmapPlot}
\alias{HeatmapPlot}
\alias{HeatmapPlot,EpivizEnvironment-method}
\title{Method to add Heatmap Plot}
\usage{
HeatmapPlot(x, y, ...)

\S4method{HeatmapPlot}{EpivizEnvironment}(x, y, ...)
}
\arguments{
\item{x}{an object of type \code{\link{EpivizEnvironment}} or \code{\link{EpivizNavigation}}}

\item{y}{a genomic data object}

\item{...}{other parameters for the plot method}
}
\description{
Method to add Heatmap Plot
}
