% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{json_parser}
\alias{json_parser}
\title{JSON parser used by this package}
\usage{
json_parser(json_str, file, method = "C", unexpected.escape = "error",
  simplify = TRUE)
}
\arguments{
\item{json_str}{json string to parse}

\item{file}{file to read json_Str from}

\item{method}{method used to parse json}

\item{unexpected.escape}{handling escape characters, one of error, skip, keep}

\item{simplify}{if TRUE, convert json-encoded lists to vectors}
}
\value{
a JSON object
}
\description{
Currently this just renames \code{\link{fromJSON}} in the \code{rjson} package.
}
\examples{
json_parser('{"a":true, "b":false, "c":null}')
}
\seealso{
\code{\link{fromJSON}}
}
