% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{knit_print.EpivizPolymer}
\alias{knit_print.EpivizPolymer}
\title{Knit print method for EpivizPolymer objects}
\usage{
knit_print.EpivizPolymer(x, ...)
}
\description{
Knit print method for EpivizPolymer objects
}
