% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{LinePlot,ANY-method}
\alias{LinePlot,ANY-method}
\title{Method to add Line Plot}
\usage{
\S4method{LinePlot}{ANY}(x, y, ...)
}
\arguments{
\item{x}{an object of type \code{\link{EpivizEnvironment}} or \code{\link{EpivizNavigation}}}

\item{y}{a genomic data object}

\item{...}{other parameters for the plot method}
}
\description{
Method to add Line Plot
}
