% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{MultiStackedLineTrack}
\alias{MultiStackedLineTrack}
\alias{MultiStackedLineTrack,EpivizEnvironment-method}
\title{Method to add Multi Stacked Line Track}
\usage{
MultiStackedLineTrack(x, y, ...)

\S4method{MultiStackedLineTrack}{EpivizEnvironment}(x, y, ...)
}
\arguments{
\item{x}{an object of type \code{\link{EpivizEnvironment}} or \code{\link{EpivizNavigation}}}

\item{y}{a genomic data object}

\item{...}{other parameters for the plot method}
}
\description{
Method to add Multi Stacked Line Track
}
