% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rand_id}
\alias{rand_id}
\title{Random ID generator for epiviz charts}
\usage{
rand_id(prefix = "")
}
\arguments{
\item{prefix}{prefix for random ID}
}
\value{
random ID
}
\description{
Random ID generator for epiviz charts
}
