% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{StackedBlocksTrack}
\alias{StackedBlocksTrack}
\alias{StackedBlocksTrack,EpivizEnvironment-method}
\title{Method to add Stacked Blocks Track}
\usage{
StackedBlocksTrack(x, y, ...)

\S4method{StackedBlocksTrack}{EpivizEnvironment}(x, y, ...)
}
\arguments{
\item{x}{an object of type \code{\link{EpivizEnvironment}} or \code{\link{EpivizNavigation}}}

\item{y}{a genomic data object}

\item{...}{other parameters for the plot method}
}
\description{
Method to add Stacked Blocks Track
}
