% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{StackedLinePlot,ANY-method}
\alias{StackedLinePlot,ANY-method}
\title{Method to add Stacked Line Plot}
\usage{
\S4method{StackedLinePlot}{ANY}(x, y, ...)
}
\arguments{
\item{x}{an object of type \code{\link{EpivizEnvironment}} or \code{\link{EpivizNavigation}}}

\item{y}{a genomic data object}

\item{...}{other parameters for the plot method}
}
\description{
Method to add Stacked Line Plot
}
