% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{StackedLineTrack}
\alias{StackedLineTrack}
\alias{StackedLineTrack,EpivizEnvironment-method}
\title{Method to add Stacked Line Track}
\usage{
StackedLineTrack(x, y, ...)

\S4method{StackedLineTrack}{EpivizEnvironment}(x, y, ...)
}
\arguments{
\item{x}{an object of type \code{\link{EpivizEnvironment}} or \code{\link{EpivizNavigation}}}

\item{y}{a genomic data object}

\item{...}{other parameters for the plot method}
}
\description{
Method to add Stacked Line Track
}
