% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_data.R
\docType{data}
\name{sumexp}
\alias{sumexp}
\title{Example SummarizedExperiment for epivizr vignette}
\format{
A \code{\link[SummarizedExperiment:RangedSummarizedExperiment]{SummarizedExperiment::RangedSummarizedExperiment}} object.
}
\usage{
data(sumexp)
}
\description{
Example SummarizedExperiment for epivizr vignette
}
