% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{TranscriptTrack}
\alias{TranscriptTrack}
\alias{TranscriptTrack,EpivizEnvironment-method}
\title{Method to add Transcript Track}
\usage{
TranscriptTrack(x, y, ...)

\S4method{TranscriptTrack}{EpivizEnvironment}(x, y, ...)
}
\arguments{
\item{x}{an object of type \code{\link{EpivizEnvironment}} or \code{\link{EpivizNavigation}}}

\item{y}{a genomic data object}

\item{...}{other parameters for the plot method}
}
\description{
Method to add Transcript Track
}
