% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_data.R
\docType{data}
\name{bcode_eset}
\alias{bcode_eset}
\title{Example expression data from the Gene Expression barcode project}
\format{An \code{\link[Biobase]{ExpressionSet}} object}
\usage{
bcode_eset
}
\value{
A \code{\link[Biobase]{ExpressionSet}} object with gene expression barcode data
}
\description{
See vignette `data_preprocessing` for code to build this object.
}
\keyword{datasets}
