% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_data.R
\docType{data}
\name{cgi_gr}
\alias{cgi_gr}
\title{Locations of CpG Islands}
\format{A \code{\link[GenomicRanges]{GRanges}} object}
\usage{
cgi_gr
}
\value{
A \code{\link[GenomicRanges]{GRanges}} object with locations of CpG Islands
}
\description{
Locations of CpG Islands
}
\keyword{datasets}
