% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chart_default_settings_colors}
\alias{chart_default_settings_colors}
\title{Get default chart settings and colors}
\usage{
chart_default_settings_colors(chart_type)
}
\arguments{
\item{chart_type}{chart type}
}
\value{
list of settings and colors
}
\description{
Get default chart settings and colors
}
