% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chart_dependencies}
\alias{chart_dependencies}
\title{HTML dependencies of an EpivizChart}
\usage{
chart_dependencies(knitr = FALSE)
}
\arguments{
\item{knitr}{whether knitr is in progress}
}
\description{
HTML dependencies of an EpivizChart
}
