% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chart_type_to_tag_name}
\alias{chart_type_to_tag_name}
\title{Get epiviz chart component name from chart type}
\usage{
chart_type_to_tag_name(chart)
}
\arguments{
\item{chart}{chart type, e.g., 'HeatmapPlot'}
}
\value{
epiviz chart component name
}
\description{
Get epiviz chart component name from chart type
}
