% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.settings_as_df}
\alias{.settings_as_df}
\title{(taken from epivizr) print settings in a readable format}
\usage{
.settings_as_df(chart_settings)
}
\description{
(taken from epivizr) print settings in a readable format
}
