% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizChart-class.R
\docType{class}
\name{EpivizChart-class}
\alias{EpivizChart-class}
\alias{EpivizChart}
\title{Data container for an epiviz chart.}
\description{
Data container for an epiviz chart.
}
\section{Fields}{

\describe{
\item{\code{data}}{Character string of an epiviz chart's data attribute.}

\item{\code{colors}}{Character string of an epiviz chart's colors attribute.}

\item{\code{settings}}{character string of an epiviz chart's settings attribute.}

\item{\code{parent}}{An EpivizEnvironment where the chart is nested or NULL}
}}

\section{Methods}{

\describe{
\item{\code{get_attributes()}}{Get attributes for rendering chart. Fields that need to be in JSON
will be converted}

\item{\code{get_colors()}}{Get chart colors}

\item{\code{get_data()}}{Get chart data}

\item{\code{get_settings()}}{Get chart settings}

\item{\code{navigate(chr, start, end)}}{Navigate a chart to a genomic location}

\item{\code{renderChart()}}{Render to html}

\item{\code{revisualize(chart_type)}}{Revisualize chart as the given chart type}

\item{\code{set_colors(colors)}}{Set chart colors}

\item{\code{set_data(data)}}{Set chart data}

\item{\code{set_settings(settings)}}{Set chart settings}
}}

