% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizChartDataMgr-class.R
\docType{class}
\name{EpivizChartDataMgr-class}
\alias{EpivizChartDataMgr-class}
\alias{EpivizChartDataMgr}
\title{Class providing data manager for epiviz charts.}
\description{
Class providing data manager for epiviz charts.
}
\section{Fields}{

\describe{
\item{\code{.ms_list}}{(environment)}

\item{\code{.ms_idCounter}}{(integer)}
}}

\section{Methods}{

\describe{
\item{\code{add_measurements(obj, datasource_name = NULL,
  datasource_obj_name = deparse(substitute(obj)), ...)}}{register measurements in data manager}

\item{\code{rm_all_measurements()}}{remove all registered measurements}

\item{\code{rm_measurements(ms_obj_or_id)}}{remove registered measurements from a given data object}
}}

