% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizEnvironment-class.R
\docType{class}
\name{EpivizEnvironment-class}
\alias{EpivizEnvironment-class}
\alias{EpivizEnvironment}
\title{Epiviz Environment Class}
\description{
Epiviz Environment Class
}
\section{Fields}{

\describe{
\item{\code{range}}{(CharacterOrNULL)}

\item{\code{initializeRegions}}{(CharacterOrNULL)}

\item{\code{children}}{A list of EpivizChart/EpivizNavigation elements.}
}}

\section{Methods}{

\describe{
\item{\code{append_child(child)}}{Append chart or navigation to environment}

\item{\code{get_initializeRegions()}}{Get initializeRegions}

\item{\code{get_range()}}{Get range}

\item{\code{remove_child(child)}}{Remove chart or navigation from environment}

\item{\code{set_initializeRegions(initReg)}}{Set initializeRegions}

\item{\code{set_range(range)}}{Set range}
}}

