% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizEnvironment-class.R
\docType{class}
\name{EpivizEnvironment-class}
\alias{EpivizEnvironment-class}
\alias{EpivizEnvironment}
\title{Epiviz Environment Class}
\description{
Epiviz Environment Class
}
\section{Fields}{

\describe{
\item{\code{range}}{(CharacterOrNULL)}

\item{\code{initializeRegions}}{(CharacterOrNULL)}

\item{\code{charts}}{A list of EpivizChart/EpivizNavigation elements.}
}}

\section{Methods}{

\describe{
\item{\code{append_chart(chart)}}{Append chart or navigation to environment}

\item{\code{get_initializeRegions()}}{Get initializeRegions}

\item{\code{get_range()}}{Get range}

\item{\code{navigate(chr = NULL, start = NULL, end = NULL)}}{Navigate environment to a genomic location}

\item{\code{remove_all_charts()}}{Remove all charts from environment}

\item{\code{remove_chart(chart)}}{Remove chart from environment}

\item{\code{set_initializeRegions(initReg)}}{Set initializeRegions}

\item{\code{set_range(range)}}{Set range}
}}

