% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizEnvironment-class.R
\docType{class}
\name{EpivizEnvironment-class}
\alias{EpivizEnvironment-class}
\alias{EpivizEnvironment}
\title{Epiviz Environment Class}
\description{
Epiviz Environment Class
}
\section{Fields}{

\describe{
\item{\code{chr}}{(character) chromosome to to display in environment plot.}

\item{\code{start}}{(integer) start location to display in environment plot.}

\item{\code{end}}{(integer) end location to to display in environment plot.}

\item{\code{data_mgr}}{An object of class \code{\link[epivizrChart]{EpivizChartDataMgr}} used to serve data to epiviz environment.}

\item{\code{epiviz_env}}{An object of class \code{htmltools}{shiny.tag} used to nest chart tags in epiviz-environment tag.}
}}

\section{Methods}{

\describe{
\item{\code{get_environment()}}{Return the epiviz-environment}

\item{\code{plot(data_object, datasource_name,
  datasource_origin_name = deparse(substitute(data_object)),
  chart_type = NULL, settings = NULL, colors = NULL, ...)}}{Return a shiny.tag representing an epiviz chart and adds it as a child of the epiviz environment tag
\describe{
\item{data_object}{GenomicRanges object to attach as chart's data}
\item{datasource_name}{Name for datasource}
\item{chart_type}{Type of chart for plot (BlocksTrack, HeatmapPlot, LinePlot,LineTrack, ScatterPlot, StackedLinePlot, StackedLineTrack)}
\item{settings}{List of settings for chart}
\item{colors}{List of colors for chart}
\item{...}{Type and columns}
}}
}}

