% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizEnvironment-class.R
\docType{class}
\name{EpivizEnvironment-class}
\alias{EpivizEnvironment-class}
\alias{EpivizEnvironment}
\title{Data container for an Epiviz environment component.}
\description{
Data container for an Epiviz environment component.
}
\section{Fields}{

\describe{
\item{\code{initializeRegions}}{(ListOrNULL) List of gene names or regions to intialize navigations.}

\item{\code{charts}}{List of class \code{\link[epivizrChart]{EpivizPolymer}} used to track nested elements.}
}}

\section{Methods}{

\describe{
\item{\code{append_chart(chart)}}{Append chart or navigation to environment}

\item{\code{get_attributes()}}{Get attributes for rendering chart}

\item{\code{get_charts()}}{Get charts within environment}

\item{\code{get_initializeRegions()}}{Get initializeRegions}

\item{\code{init_region(chr = NULL, start = NULL, end = NULL)}}{Initialize navigation based on a genomic region
\describe{
  \item{chr}{Chromosome}
  \item{chr}{Start location}
  \item{chr}{End location}
}}

\item{\code{init_regions(regions)}}{Initialize navigations based on a genomic regions
\describe{
  \item{regions}{List of named lists of genomic locations, e.g.,
  list(list(chr='chr11', start=99800000, end=103383180))}
}}

\item{\code{navigate(chr = NULL, start = NULL, end = NULL)}}{Navigate environment to genomic location
\describe{
  \item{chr}{Chromosome}
  \item{start}{Start location}
  \item{end}{End location}
}}

\item{\code{order_charts(ordered_charts)}}{Order the charts within an environment
\describe{
\item{charts}{An ordered list of EpivizPolymer objects}
}}

\item{\code{remove_all_charts()}}{Remove all charts from environment}

\item{\code{remove_chart(chart)}}{Remove chart from environment}

\item{\code{renderChart()}}{Render to html}

\item{\code{set_initializeRegions(initReg)}}{Set initializeRegions}
}}

