% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizEnvironment-class.R
\docType{class}
\name{EpivizEnvironment-class}
\alias{EpivizEnvironment-class}
\alias{EpivizEnvironment}
\title{Data container for an Epiviz environment component.}
\description{
Data container for an Epiviz environment component.
}
\section{Fields}{

\describe{
\item{\code{initializeRegions}}{(ListOrNULL) List of gene names or regions to intialize navigations.}

\item{\code{charts}}{List of class \code{\link[epivizrChart]{EpivizPolymer}} used to track nested elements.}
}}

\section{Methods}{

\describe{
\item{\code{append_chart(chart)}}{Append chart or navigation to environment}

\item{\code{get_attributes()}}{Get attributes for rendering chart}

\item{\code{get_initializeRegions()}}{Get initializeRegions}

\item{\code{navigate(chr = NULL, start = NULL, end = NULL)}}{Navigate environment to genomic location
\describe{
  \item{chr}{{Chromosome}
  \item{start}{{Start location}
  \item{end}{{End location}
}}

\item{\code{remove_all_charts()}}{Remove all charts from environment}

\item{\code{remove_chart(chart)}}{Remove chart from environment}

\item{\code{renderChart()}}{Render to html}

\item{\code{set_initializeRegions(initReg)}}{Set initializeRegions}
}}

