% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epivizEnvironment.R
\name{epivizEnvironment}
\alias{epivizEnvironment}
\title{Create \code{\link{EpivizChart}} object to knit epiviz charts.}
\usage{
epivizEnvironment(chr = "chr11", start = 99800000, end = 103383180)
}
\arguments{
\item{chr}{(character) chromosome to browse to on app startup.}

\item{start}{(integer) start location to browse to on app startup.}

\item{end}{(integer) end location to browse to on app startup.}

\item{data_mgr}{(EpivizChartDataMgr)}
}
\value{
An object of class \code{\link{EpivizChart}}
}
\description{
Create \code{\link{EpivizChart}} object to knit epiviz charts.
}
\examples{
# see package vignete for example usage
epiviz <- epivizEnvironment(chr="chr11", start=99800000, end=103383180)

}
