% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizCharts-classes.R
\docType{class}
\name{EpivizHeatmapPlot-class}
\alias{EpivizHeatmapPlot-class}
\alias{EpivizHeatmapPlot}
\title{Data container for an Epiviz Heatmap Plot.}
\description{
Data container for an Epiviz Heatmap Plot.
}

