% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizNavigation-class.R
\docType{class}
\name{EpivizNavigation-class}
\alias{EpivizNavigation-class}
\alias{EpivizNavigation}
\title{Data container for an Epiviz navigation component.}
\description{
Data container for an Epiviz navigation component.
}
\section{Fields}{

\describe{
\item{\code{gene}}{(CharacterOrNULL) Gene}

\item{\code{geneInRange}}{(CharacterOrNULL) Nearest Gene in range.}

\item{\code{parent}}{An object of class \code{\link[epivizrChart]{EpivizEnvironment}} where navigation is appended.}
}}

\section{Methods}{

\describe{
\item{\code{get_attributes()}}{Get attributes for rendering chart}

\item{\code{get_gene()}}{Get gene}

\item{\code{get_geneInRange()}}{Get gene in range}

\item{\code{set_gene(gene)}}{Set gene}

\item{\code{set_geneInRange(gene)}}{Set step ratio}
}}

