% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizNavigation-class.R
\docType{class}
\name{EpivizNavigation-class}
\alias{EpivizNavigation-class}
\alias{EpivizNavigation}
\title{Epiviz Navigation Class}
\description{
Epiviz Navigation Class
}
\section{Fields}{

\describe{
\item{\code{chr}}{(character) chromosome to to display in environment plot.}

\item{\code{start}}{(numeric) start location to display in environment plot}

\item{\code{end}}{(numeric) end location to to display in environment plot.}

\item{\code{gene}}{(CharacterOrNULL)}

\item{\code{strRange}}{(NumericOrNULL)}

\item{\code{stepRatio}}{(NumericOrNULL)}

\item{\code{zoomRatio}}{(NumericOrNULL)}

\item{\code{collapsed}}{(LogicalOrNULL)}

\item{\code{geneInRange}}{(CharacterOrNULL)}

\item{\code{configSrc}}{(CharacterOrNULL)}
}}

\section{Methods}{

\describe{
\item{\code{get_collapsed()}}{Get collapsed}

\item{\code{get_configSrc(src)}}{Get config}

\item{\code{get_gene()}}{Get gene}

\item{\code{get_geneInRange()}}{Get gene in range}

\item{\code{get_stepRatio()}}{Get step ratio}

\item{\code{get_strRange()}}{Get range}

\item{\code{get_zoomRatio()}}{Get step ratio}

\item{\code{set_collapsed(collapse)}}{Set step ratio}

\item{\code{set_configSrc(src)}}{Set step ratio}

\item{\code{set_gene(gene)}}{Set gene}

\item{\code{set_geneInRange(gene)}}{Set step ratio}

\item{\code{set_stepRatio(ratio)}}{Set step ratio}

\item{\code{set_strRange(range)}}{Set range}

\item{\code{set_zoomRatio(ratio)}}{Set step ratio}
}}

