% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizNavigation-class.R
\docType{class}
\name{EpivizNavigation-class}
\alias{EpivizNavigation-class}
\alias{EpivizNavigation}
\title{Data container for an Epiviz navigation component.}
\description{
Data container for an Epiviz navigation component.
}
\section{Fields}{

\describe{
\item{\code{gene}}{(character) Gene}

\item{\code{geneInRange}}{(character) Nearest Gene in range.}

\item{\code{parent}}{An object of class \code{\link[epivizrChart]{EpivizEnvironment}} where navigation is appended.}
}}

\section{Methods}{

\describe{
\item{\code{clone_charts(charts)}}{Clone EpivizCharts and append to navigation. Each chart must already
exist in the navigation's data manager, otherwise an error will occur
when attempting to intialize using their measurements
\describe{
\item{charts}{list of EpivizCharts whose data exists in the
navigation's data manager }}}

\item{\code{get_attributes()}}{Get attributes for rendering chart}

\item{\code{get_gene()}}{Get gene}

\item{\code{get_geneInRange()}}{Get gene in range}

\item{\code{get_name()}}{Get name of Epiviz Web Component}

\item{\code{set_gene(gene)}}{Set gene}

\item{\code{set_geneInRange(gene)}}{Set step ratio}
}}

