% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizNavigation-class.R
\docType{class}
\name{EpivizNavigation-class}
\alias{EpivizNavigation-class}
\alias{EpivizNavigation}
\title{Epiviz Navigation Class}
\description{
Epiviz Navigation Class
}
\section{Methods}{

\describe{
\item{\code{add(polymer_obj)}}{Add chart or navigation to environment}
}}

