% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizPolymer-class.R
\docType{class}
\name{EpivizPolymer-class}
\alias{EpivizPolymer-class}
\alias{EpivizPolymer}
\title{Class for knitting epiviz charts}
\description{
Class for knitting epiviz charts
}
\section{Fields}{

\describe{
\item{\code{tag}}{An object of class \code{\link[htmltools]{shiny.tag}} representing an epiviz chart in html.}
}}


