% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizPolymer-class.R
\docType{class}
\name{EpivizPolymer-class}
\alias{EpivizPolymer-class}
\alias{EpivizPolymer}
\title{Data container for an Epiviz web component.}
\description{
Data container for an Epiviz web component.
}
\section{Fields}{

\describe{
\item{\code{data_mgr}}{\code{\link[epivizrChart]{EpivizChartDataMgr}}}

\item{\code{name}}{(character) Epiviz chart type (tag name).}

\item{\code{class}}{(CharacterOrNULL) Epiviz chart's class attribute.}

\item{\code{id}}{(CharacterOrNULL) Epiviz chart's id attribute.}

\item{\code{chr}}{(CharacterOrNULL) Chromosome location.}

\item{\code{start}}{(NumericOrNULL) Start location.}

\item{\code{end}}{(NumericOrNULL) End location.}

\item{\code{measurements}}{(ListOrNULL) list of measurements of class \code{\link[epivizrData]{EpivizMeasurement}}.}
}}

\section{Methods}{

\describe{
\item{\code{get_attributes()}}{Get attributes for rendering chart}

\item{\code{get_chr()}}{Get chromosome}

\item{\code{get_class()}}{Get class}

\item{\code{get_data_mgr()}}{Get data manager}

\item{\code{get_end()}}{Get end}

\item{\code{get_id()}}{Get id}

\item{\code{get_measurements()}}{Get measurements}

\item{\code{get_name()}}{Get name}

\item{\code{get_start()}}{Get start}

\item{\code{set_chr(chr)}}{Set the chromosome}

\item{\code{set_class(class)}}{Set chart class}

\item{\code{set_end(end)}}{Set end}

\item{\code{set_id(id)}}{Set chart id}

\item{\code{set_measurements(ms)}}{Set measurements}

\item{\code{set_name(name)}}{Set name}

\item{\code{set_start(start)}}{Set start}
}}

