% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizPolymer-class.R
\docType{class}
\name{EpivizPolymer-class}
\alias{EpivizPolymer-class}
\alias{EpivizPolymer}
\title{Class for knitting epiviz charts}
\description{
Class for knitting epiviz charts
}
\section{Fields}{

\describe{
\item{\code{chr}}{(character) chromosome to to display in environment plot.}

\item{\code{start}}{(integer) start location to display in environment plot.}

\item{\code{end}}{(integer) end location to to display in environment plot.}

\item{\code{data_mgr}}{An object of class \code{\link[epivizrData]{EpivizDataMgr}} used to serve data to epiviz charts.}

\item{\code{epiviz_envir}}{An object of class \code{shiny.tag} used to nest chart tags in epiviz-environment tag.}
}}

\section{Methods}{

\describe{
\item{\code{plot(data_object, datasource_name = deparse(substitute(data_object)),
  settings = NULL, colors = NULL, ...)}}{Return a shiny.tag representing an epiviz chart and appends to the epiviz environment
\describe{
\item{data_object}{GenomicRanges object to attach as chart's data}
\item{datasource_name}{Name for datasource}
\item{settings}{List of settings for chart}
\item{colors}{List of colors for chart}
\item{...}{Type and columns}
}}
}}

