% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizPolymer-class.R
\docType{class}
\name{EpivizPolymer-class}
\alias{EpivizPolymer-class}
\alias{EpivizPolymer}
\title{Parent data container for an epiviz chart.}
\description{
Parent data container for an epiviz chart.
}
\section{Fields}{

\describe{
\item{\code{data_mgr}}{EpivizChartDataMgr.}

\item{\code{name}}{Character string representing an epiviz chart type (tag name).}

\item{\code{class}}{Character string of an epiviz chart's class attribute.}

\item{\code{id}}{Character string of an epiviz chart's id attribute.}

\item{\code{chr}}{(CharacterOrNULL) chromosome to to display in environment plot.}

\item{\code{start}}{(NumericOrNULL) start location to display in environment plot.}

\item{\code{end}}{(NumericOrNULL) end location to to display in environment plot.}

\item{\code{measurements}}{Character string of an epiviz chart's measurements.}
}}

\section{Methods}{

\describe{
\item{\code{get_attributes()}}{Get attributes for rendering chart. Fields that need to be in JSON
will be converted}

\item{\code{get_chr()}}{Get chromosome}

\item{\code{get_class()}}{Get class}

\item{\code{get_data_mgr()}}{Return EpivizChart Data Manager}

\item{\code{get_end()}}{Get end}

\item{\code{get_id()}}{Get id}

\item{\code{get_measurements()}}{Get measurements}

\item{\code{get_name()}}{Get name}

\item{\code{get_start()}}{Get start}

\item{\code{set_chr(chr)}}{Set the chromosome}

\item{\code{set_class(class)}}{Set chart class}

\item{\code{set_end(end)}}{Set end}

\item{\code{set_id(id)}}{Set chart id}

\item{\code{set_measurements(ms)}}{Set measurements}

\item{\code{set_name(name)}}{Set name}

\item{\code{set_start(start)}}{Set start}
}}

