% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizCharts-classes.R
\docType{class}
\name{EpivizScatterPlot-class}
\alias{EpivizScatterPlot-class}
\alias{EpivizScatterPlot}
\title{Data container for an Epiviz Scatter Plot.}
\description{
Data container for an Epiviz Scatter Plot.
}
\section{Methods}{

\describe{
\item{\code{get_chart_type()}}{Get chart type}

\item{\code{get_default_settings()}}{Get default settings}

\item{\code{get_name()}}{Get name}
}}

