% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizCharts-classes.R
\docType{class}
\name{EpivizStackedLineTrack-class}
\alias{EpivizStackedLineTrack-class}
\alias{EpivizStackedLineTrack}
\title{Data container for an Epiviz Stacked Line Track.}
\description{
Data container for an Epiviz Stacked Line Track.
}

