% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_data.R
\docType{data}
\name{sumexp}
\alias{sumexp}
\title{Example SummarizedExperiment for epivizr vignette}
\format{An object of class \code{RangedSummarizedExperiment} with 12800 rows and 2 columns.}
\usage{
sumexp
}
\value{
A \code{\link[SummarizedExperiment]{RangedSummarizedExperiment}} object.
}
\description{
Example SummarizedExperiment for epivizr vignette
}
\keyword{datasets}
